requirements.txt
asgiref==3.6.0
Django==4.2
djangorestframework==3.14.0
et-xmlfile==1.1.0
mysqlclient==2.1.1
numpy==1.24.3
openpyxl==3.1.2
pandas==2.0.1
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3

requests~=2.29.0
pip~=21.3.1
wheel~=0.37.1
setuptools~=60.2.0
crontab~=1.0.1
certifi~=2022.12.7
idna~=3.4
urllib3~=1.26.15